create fetch data in products component
