+++ prisma/seed.js
@@ -6,6 +6,7 @@ const plants = [
   { id: 2, image_link:'/products/pot_02.jpg', name: 'sun flower', price:8, description: 'Sunflowers are bee magnets; each flower head produces a large volume of pollen and nectar that lures in bees and other pollinators'},
   { id: 3, image_link:'/products/pot_03.jpg', name: 'Croton', price:12, description: 'Native to India, Sri Lanka, Malaysia and the western Pacific islands, the croton enjoys hot, humid climates.'},
   { id: 4, image_link:'/products/pot_04.jpg', name: 'Aspidistra', price:15, description: 'Aspidistra are sometimes known as the Cast Iron Plant because they’re almost impossible to kill. They’re practically bomb-proof (we have not tested this). '},
+  { id: 5, image_link:'/products/pot_05.jpg', name: 'Aspidistra2', price:15, description: 'Aspidistra are sometimes known as the Cast Iron Plant because they’re almost impossible to kill. They’re practically bomb-proof (we have not tested this). '},
 ]
 
 async function main() {
